Feature/book

+++ src/main/java/shop/itbook/itbookshop/productgroup/product/entity/Product.java
@@ -16,7 +16,8 @@ import lombok.Setter;
 /**
  * 상품에 대한 엔티티 입니다.
  *
- * @author 노수연 * @since 1.0
+ * @author 노수연
+ * @author 이하늬
  * @since 1.0
  */
 @Getter
@@ -67,7 +68,7 @@ public class Product {
     private Integer increasePointPercent;
 
     @Column(name = "discount_percent", nullable = false, columnDefinition = "integer default 0")
-    private Integer discountPercent;
+    private Double discountPercent;
 
     @Column(name = "raw_price", nullable = false)
     private Long rawPrice;
@@ -88,14 +89,14 @@ public class Product {
      * @param increasePointPercent 상품 포인트 적립율입니다.
      * @param discountPercent      상품 할인율입니다.
      * @param rawPrice             상품 매입원가입니다.
-     * @author
+     * @author 이하늬
      */
     @SuppressWarnings("java:S107") // 생성자 필드 갯수가 많아 추가
     @Builder
     public Product(String name, String simpleDescription, String detailsDescription, Integer stock,
                    LocalDateTime productCreatedAt, Boolean isSelled, Boolean isDeleted,
                    String thumbnailUrl, Long dailyHits, Long fixedPrice,
-                   Integer increasePointPercent, Integer discountPercent, Long rawPrice) {
+                   Integer increasePointPercent, Double discountPercent, Long rawPrice) {
         this.name = name;
         this.simpleDescription = simpleDescription;
         this.detailsDescription = detailsDescription;
